CUDASetup: improve DLL finding code
Added a cache variable check, so that we only try the file glob
once, instead of repeating it for every CMake run.

Added another glob expression to check - exact name with dll suffix.
No current libs use that pattern, but it's a pretty common one that
might be encountered in future versions of the toolkit.