AddFlagsHelpers: get list of enabled languages inside functions
Before, was just pulling the list once and sticking it in the
'enabled_languages' variable, then referencing this variable in
each function.

This was bad, because it's assuming that the user won't modify
the contents of enable_languages in between the point where the
module is included, and the point where one of the helper functions
is called. Since we can't really guarantee this, it's better just
to recheck it for every function call.