InstallDeps: added INSTALL_DEPS_AUTO_MODE option
Can be set to NONE, LIMITED, or ALL. LIMITED is the default.

This variable determines if/how implicit dependencies of the explicit
libs passed into install_deps() are installed.

Before this commit, the module behaved as if it was in ALL mode.
So, in addition to adding the option, we're effectively changing
the default behavior from ALL to LIMITED, as well.